@@ -1,8 +1,9 @@
-package com.smartmobe.photoalbummvvm.util;
+package com.sharmadhiraj.photoalbummvvm.util;
 
 import android.content.Context;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by Dhiraj Sharma on April 06, 2017
@@ -11,7 +12,7 @@ import android.support.v7.app.AppCompatActivity;
 public class ViewUtil {
 
     public static void displayHomeAsUpEnabled(Context context) {
-        ActionBar actionBar = ((AppCompatActivity)context).getSupportActionBar();
+        ActionBar actionBar = ((AppCompatActivity) context).getSupportActionBar();
         if (actionBar != null) {
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
