@@ -1,19 +1,19 @@
-package com.smartmobe.photoalbummvvm.view.activity;
+package com.sharmadhiraj.photoalbummvvm.view.activity;
 
 import android.content.Context;
 import android.content.Intent;
-import android.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
 import android.view.MenuItem;
 
-import com.smartmobe.photoalbummvvm.R;
-import com.smartmobe.photoalbummvvm.databinding.PhotoDetailActivityBinding;
-import com.smartmobe.photoalbummvvm.model.Photo;
-import com.smartmobe.photoalbummvvm.util.ViewUtil;
-import com.smartmobe.photoalbummvvm.viewmodel.PhotoDetailViewModel;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.databinding.DataBindingUtil;
+
+import com.sharmadhiraj.photoalbummvvm.R;
+import com.sharmadhiraj.photoalbummvvm.databinding.PhotoDetailActivityBinding;
+import com.sharmadhiraj.photoalbummvvm.model.Photo;
+import com.sharmadhiraj.photoalbummvvm.util.ViewUtil;
+import com.sharmadhiraj.photoalbummvvm.viewmodel.PhotoDetailViewModel;
 
 /**
  * Created by Dhiraj Sharma on April 05, 2017
@@ -39,37 +39,19 @@ public class PhotoDetailActivity extends AppCompatActivity {
     }
 
 
-
     private void getExtrasFromIntent() {
         Photo photo = (Photo) getIntent().getSerializableExtra(EXTRA_PHOTO);
         PhotoDetailViewModel photoDetailViewModel = new PhotoDetailViewModel(photo);
         photoDetailActivityBinding.setPhotoDetailViewModel(photoDetailViewModel);
-        setTitle(photo.getTitle());
+        if (photo != null) setTitle(photo.getTitle());
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                finish();
-                break;
-            case R.id.menu_comment:
-                startActivityActionView();
-                break;
+        if (item.getItemId() == android.R.id.home) {
+            finish();
         }
         return super.onOptionsItemSelected(item);
-
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.comment, menu);
-        return true;
-    }
-
-    private void startActivityActionView() {
-        startActivity(
-                new Intent(this, CommentActivity.class));
     }
 
 }
